requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sentence-transformers
llama-cpp-python
scikit-learn
numpy
pdfplumber
beautifulsoup4
PyMuPDF
faiss-cpu